Create the concept of a serial_buffer and implement it
We have this odd DeviceConfig struct, which used a poor description
and had no usable methods.  This change replaces the concept of
DeviceConfig with a `struct serial_buffer`.  This is a serial device
with an associated buffer for reading in data from the USART.  It also
has some useful methods associated with it and helps clean up a bit of
a mess.