Improve Serial Logging Abilities
To be able to log cellular data well, methods that put various values
to the cellular serial lines were wrapped.  This sucks because it
creates bloat and make code harder to maintain.  To fix this, this
change enables one to set a callback in the Serial struct.  By setting
this callback, a method can be invoked every time data is put on the
serial bus of the given serial object.  By doing this we avoid the
duplicate code and still have the ability to get the behavior we want.